+++ frontend/src/utils/getRandomNumber.ts
@@ -1,0 +1,1 @@
+export const getRandomNumber = () => Math.floor(Math.random() * 3);
